backend/requirements.txt
# Backend-specific dependencies for Lecture Transcriber API
# These are the minimal dependencies needed to run the FastAPI backend

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# File handling
python-multipart>=0.0.6  # For file uploads
aiofiles>=23.2.0         # Async file operations

# Template engine
jinja2>=3.1.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Core processing dependencies (from main requirements.txt)
# These are needed for the processing pipeline
ffmpeg-python>=0.2.0
librosa>=0.10.0
transformers>=4.30.0
torch>=2.0.0
torchaudio>=2.0.0
huggingface-hub>=0.16.0
numpy>=1.24.0

# Utilities
rich>=13.0.0
tqdm>=4.65.0

# Note: For full functionality, install the main requirements.txt at project root
# pip install -r ../requirements.txt
